tests/wpt/tests/html/canvas/offscreen/text/2d.text.measure.advances-not-defined.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>OffscreenCanvas test: 2d.text.measure.advances-not-defined</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.text.measure.advances-not-defined</h1>
<p class="desc">Testing that TextMetrics.advances is not defined</p>


<script>
promise_test(async t => {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.font = '50px CanvasTest';
  _assert(!("advances" in ctx.measureText('Hello')), "!(\"advances\" in ctx.measureText('Hello'))");

}, "Testing that TextMetrics.advances is not defined");
</script>
